controllers: move some APIs into the tiltfile reconciler package
